requirements.txt
langchain
pydantic
langgraph
langchain_core
langchain-groq
python-dotenv
requests
streamlit
rich